Move RPSCount to a separate file
Fixes #180
