+++ ndbench-core/src/main/java/com/netflix/ndbench/core/NdBenchDriver.java
@@ -45,7 +45,6 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicReference;
 
 /**
@@ -464,66 +463,4 @@ public class NdBenchDriver {
         return keyGeneratorReadRef.get();
     }
 
-    static class RPSCount {
-        private final AtomicLong reads = new AtomicLong(0L);
-        private final AtomicLong writes = new AtomicLong(0L);
-        private final IConfiguration config;
-        private final NdBenchMonitor ndBenchMonitor;
-        private final AtomicReference<RateLimiter> readLimiter;
-        private final AtomicReference<RateLimiter> writeLimiter;
-        private final AtomicBoolean readsStarted;
-        private final AtomicBoolean writesStarted;
-
-        RPSCount(AtomicBoolean readsStarted,
-                 AtomicBoolean writesStarted,
-                 AtomicReference<RateLimiter> readLimiter,
-                 AtomicReference<RateLimiter> writeLimiter,
-                 IConfiguration config,
-                 NdBenchMonitor ndBenchMonitor) {
-
-            this.readsStarted = readsStarted;
-            this.writesStarted = writesStarted;
-            this.readLimiter = readLimiter;
-            this.writeLimiter = writeLimiter;
-            this.config = config;
-            this.ndBenchMonitor = ndBenchMonitor;
-        }
-
-
-        void updateRPS() {
-            int secondsFreq = config.getStatsUpdateFreqSeconds();
-
-
-            long totalReads = ndBenchMonitor.getReadSuccess() + ndBenchMonitor.getReadFailure();
-            long totalWrites = ndBenchMonitor.getWriteSuccess() + ndBenchMonitor.getWriteFailure();
-            long totalOps = totalReads + totalWrites;
-            long totalSuccess = ndBenchMonitor.getReadSuccess() + ndBenchMonitor.getWriteSuccess();
-
-            long readRps = (totalReads - reads.get()) / secondsFreq;
-            long writeRps = (totalWrites - writes.get()) / secondsFreq;
-
-            long sRatio = (totalOps > 0) ? (totalSuccess * 100L / (totalOps)) : 0;
-
-            reads.set(totalReads);
-            writes.set(totalWrites);
-            ndBenchMonitor.setWriteRPS(writeRps);
-            ndBenchMonitor.setReadRPS(readRps);
-
-            logger.info("Read avg: "  + (double) ndBenchMonitor.getReadLatAvg() / 1000.0  + "ms, Read RPS: "  + readRps
-                    + ", Write avg: " + (double) ndBenchMonitor.getWriteLatAvg() / 1000.0 + "ms, Write RPS: " + writeRps
-                    + ", total RPS: " + (readRps + writeRps) + ", Success Ratio: " + sRatio + "%");
-            long expectedReadRate = (long) this.readLimiter.get().getRate();
-            long expectedwriteRate = (long) this.writeLimiter.get().getRate();
-            String bottleneckMsg = "If this occurs consistently the benchmark client could be the bottleneck.";
-
-            if (this.config.isReadEnabled() && readsStarted.get() && readRps < expectedReadRate) {
-                logger.warn("Observed Read RPS ({}) less than expected read rate + ({}).\n{}",
-                        readRps, expectedReadRate, bottleneckMsg);
-            }
-            if (this.config.isWriteEnabled() && writesStarted.get() && writeRps < expectedwriteRate) {
-                logger.warn("Observed Write RPS ({}) less than expected write rate + ({}).\n{}",
-                        writeRps, expectedwriteRate, bottleneckMsg);
-            }
-        }
-    }
 }
